% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{calc_nb}
\alias{calc_nb}
\title{calc neighbors.}
\usage{
calc_nb(x, fun = c("dist", "knn", "poly", "poly+dist", "euclidean"), opt)
}
\arguments{
\item{x}{sf class data}

\item{fun}{Specifies the method used to calculate the adjacency points.
\itemize{
\item dist: use spdep::dnearneigh function.
\item knn: use spdep::knearneigh and spdep::knn2nb function.
\item poly: use spdep::poly2nb function.
\item eublidean: use spdep::dnearneigh function and x converted sf to data.frame.
}}

\item{opt}{set function options.}
}
\description{
calc neighbors.
}
