% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{calc_stdbscan}
\alias{calc_stdbscan}
\title{calc ST-DBSCAN clustering and return cluster label.}
\usage{
calc_stdbscan(
  nb = NULL,
  vals = setVals(),
  dbscantype = c("default", "grid"),
  minPts = 5,
  FUN = mean,
  na_value = -9999
)
}
\arguments{
\item{nb}{neighbor list}

\item{vals}{as delta_eps list}

\item{dbscantype}{\itemize{
\item "grid": For Grid data. Adjust nb size before calc.
\item "default": For normal(random point) data. not Adjustment.
}}

\item{minPts}{minimum number of points}

\item{FUN}{集計用の計算関数を指定します。}

\item{na_value}{元データにNA値が含まれている場合に代替する値を指定します。
なお、データは代替して計算されますが、NA値のクラスタリング結果は Noise 判定となります。
Noise判定にしたくない場合はこの機能を利用せず、引数指定時にNA値を除去または代替したデータを設定するようにしてください。}
}
\description{
ST-DBSCAN法でクラスタリングを行います。
https://www.sciencedirect.com/science/article/pii/S0169023X06000218
}
\details{
article{birant2007st,
title={ST-DBSCAN: An algorithm for clustering spatial--temporal data},
author={Birant, Derya and Kut, Alp},
journal={Data \& knowledge engineering},
volume={60},
number={1},
pages={208--221},
year={2007},
publisher={Elsevier}
}
}
\note{
eps1 <- 144 ## km
eps2 <- 3600 * 6
minPts <- 20
FUN <- mean
num <- 10000
vals <- setVals("test1", rnorm(num, mean = 10, sd = 1), 0.05,
"test2", matrix(rnorm(num, mean = 100, sd = 10), ncol = 10), 5)
\subsection{ramdom point and non temporal data.}{

x <- sf::st_as_sf(data.frame(lon = runif(num, min = 130, max = 140),
lat = runif(num, min = 30, max = 40)),
coords = c("lon", "lat"), crs = 4326) |>
sf::st_geometry()
nb <- stnb(x = x, eps1 = eps1, eps2 = eps2, time = NULL, neighbortype = "random")
cl <- calc_stdbscan(nb = nb$id, vals = vals, dbscantype = "default", minPts = minPts)
}
}
\keyword{internal}
