% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-STclustering.R
\name{checkCond}
\alias{checkCond}
\title{check list for Spatio-Temporal data.}
\usage{
checkCond(
  ts = list(upper = "rsq >= 0.7 & 6 >= rmse", middle =
    "0.7 > rsq & rsq >= 0.4 & 9 >= rmse", lower = "0.4 > rsq & rsq > 0.2 & 12 >= rmse"),
  sp = list(upper = "z_value > 0 & 0.01 >= adj_p_value", middle =
    "z_value > 0 & 0.05 >= adj_p_value", lower = "0.05 >= adj_p_value")
)
}
\arguments{
\item{ts}{\code{sparseDFM::sparseDFM} 関数で計算された結果について、R ^ 2値 (rsq) 及びRMSE (rmse) 値の基準を設定します。
\code{upper}, \code{middle}, \code{lower} の3段階の設定ができます。}

\item{sp}{\code{spdep::joincount.mc} 及び \code{spdep::joincount.multi} 関数で計算された結果について、z値 (z_value) 及び調整済みp値 (adj_p_value) の基準を設定します。
\code{upper}, \code{middle}, \code{lower} の3段階の設定ができます。}
}
\description{
時系列データの類似性及び空間データの集塊性の確認の基準を設定します。
}
