% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-STclustering.R
\name{createBufferMap}
\alias{createBufferMap}
\title{Creates a specific area based on the location of the measuring station.}
\usage{
createBufferMap(
  geo = sf::st_sfc(),
  voronoi = TRUE,
  buffer = TRUE,
  dist = 10000,
  coast = TRUE,
  coastline = sf::sfc()
)
}
\arguments{
\item{geo}{sfcクラスの位置情報を指定します。}

\item{voronoi}{TRUE の場合、ボロノイ分割地図を作成します。}

\item{buffer}{TRUE の場合、測定局中心地点から \code{dist} 引数で指定する距離のバッファー領域を作成します。
\code{voronoi = TRUE} の場合は、ボロノイ分割地図とバッファー領域が重なる領域 (intersection) を出力します。}

\item{dist}{\code{buffer = TRUE} の場合に作成するバッファー領域半径 (m) を指定します。}

\item{coast}{TRUE の場合、海岸線で全体をクリッピングします。}

\item{coastline}{\code{coast = TRUE} の場合に利用する海岸線情報をsfcクラスのポリゴンで指定します。}
}
\description{
測定局の位置情報から、バッファー領域、ボロノイ分割（測定局間の距離が等しくなるライン）を作成し、海岸線でクリッピングします。
また、作成されたこれらの地図の重なりでクリッピングを行います。
}
\examples{
createBufferMap(geo = geo, voronoi = TRUE, buffer = TRUE, dist = 20000, coast = TRUE, coastline = sf::st_cfc())
}
