% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstdbscan.R
\name{cutclust}
\alias{cutclust}
\title{\code{hstdbscan} 関数で計算したクラスターの結果を分割します。}
\usage{
cutclust(x, k, ...)
}
\arguments{
\item{x}{\code{hstdbscan} の結果オブジェクトを指定します。}

\item{k}{クラスターの分割数を整数で指定します。}

\item{...}{その他、 \code{cutree} 関数で使用できる引数を指定することができます。}
}
\value{
階層的クラスタリング結果のラベル、位置情報が含まれる data.frame が返ります。
}
\description{
\code{hstdbscan} 関数で計算したクラスターの結果を分割します。
}
\examples{
eps1 <- 144 ## km
eps2 <- 3600 * 6
time <- NULL
minPts <- 4
FUN <- mean
time <- as.POSIXct("2024-01-01 00:00:00", tz = "JST") + 3600 * seq(0, 23, by = 6)
x <- sf::st_as_sf(expand.grid(seq(130, 140, by = 1), seq(30, 40, by = 1)),
                  coords = c("Var1", "Var2"), crs = 4326) |> sf::st_geometry()
vals <- setVals("test1", rnorm(length(x) * length(time), mean = 10, sd = 1), 0.5,
                "test2", matrix(rnorm(length(x) * length(time), mean = 100, sd = 10), ncol = length(time)), 5)
clust <- hstdbscan(x = x, time = time, eps1 = eps1, eps2 = eps2, minPts = 4,
                   vals = vals, neighbortype = "spatial")
cutclust(clust, 4)
}
