% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCluster.R
\name{geostclust}
\alias{geostclust}
\title{run any clustering method.}
\usage{
geostclust(
  x = double(),
  geo,
  times,
  cuts = c(4, 8, 16, 32),
  method = c("ward", "kmeans", "kshape", "dtw", "hstdbscan", "others"),
  ...
)
}
\arguments{
\item{x}{行方向がユニークな位置、列方向が時系列となる2次元配列のデータを指定します。}

\item{geo}{位置情報を \code{sf} または \code{sfc} 形式で指定します。}

\item{times}{時間情報を指定します。}

\item{cuts}{クラスター分割数を整数で指定します。複数同時に指定する場合はベクトル (\code{c(1, 2, 3, ...)} のように指定してください。)}

\item{method}{クラスタリング方法を指定します。ウォード法 ("ward"), k-means法 ("kmeans"), k-shape法 ("kshape"), 動的時間伸縮法 ("dtw"), 階層的ST-DBSCAN法 ("hstdbscan") から一つ指定します。}

\item{...}{クラスタリングの引数オプションを指定します。}
}
\value{
\itemize{
\item cluster
cuts で指定した分割数 (k) 毎にクラスタリングラベルが作成され、位置情報とマージされて sf 形式のデータで格納されます。
\item model
クラスタリングを実行した結果が格納されています。 \code{type = "partitonal"} の場合は、cuts 毎にモデルが作成され、リスト形式で格納されています。
\item type
クラスタリングの方法が階層的 ("hierarchical") か、非階層的 ("partitional") か示されます。
\item method
クラスタリング方法で指定した method 引数の値が格納されます。
}
}
\description{
複数のクラスタリング方法でクラスター解析を実施します。
}
\examples{
x <- seq(130, 140, by = 1)
y <- seq(30, 40, by = 1)
t <- as.POSIXct("2024-01-01 00:00:00", tz = "JST") + 3600 * seq(0, 23, by = 6)
geo <- sf::st_as_sf(expand.grid(x, y), coords = c("Var1", "Var2"), crs = 4326)
D <- abs(runif(nrow(geo) * length(t))) * 100
D <- Reduce(cbind, split(D, do.call(c, Map(rep, t, nrow(geo)))), init = NULL)
clust <- geostclust(x = D, geo = geo, times = t, method = "ward")
clust <- geostclust(x = D, geo = geo, times = t, method = "kmeans")
clust <- geostclust(x = D, geo = geo, times = t, method = "kshape")
clust <- geostclust(x = D, geo = geo, times = t, method = "dtw")
clust <- geostclust(geo = geo, times = t, method = "hstdbscan", eps1 = 144, eps2 = 3600 * 6, minPts = 6,
                    vals = list(list(D = D,
                                     delta_eps = 20)))
}
