% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-STclustering.R
\name{get_num_factor}
\alias{get_num_factor}
\title{get num of sparseDFM model's factor.}
\usage{
get_num_factor(x)
}
\arguments{
\item{x}{set ts-data}
}
\description{
get num of sparseDFM model's factor.
}
