% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{getVals}
\alias{getVals}
\title{Obtains data at an arbitrary time position from the data created by the setVals function.}
\usage{
getVals(x, timelength = NULL, time = NULL, pos)
}
\arguments{
\item{x}{setVals関数で作成されたデータを指定します。}

\item{timelength}{各時間のデータが同数の場合、全時間の長さを指定します。}

\item{time}{各時間のデータが同数ではない場合、setVals関数のv変数と同数の時間情報ベクトルを指定します}

\item{pos}{抽出する時間の位置または時間の値を指定します。
timelengthに値を指定した場合は、何番目の時間のデータを抽出するか指定します。このとき、x$vのデータはncol = timelength のmatrixに変換され、pos 列目のデータを返すようになります。(as.matrix(x$v, ncol = timelength)\link{, pos})
time引数にデータを指定した場合は、抽出する時間値そのものを指定します。データに一致する情報だけを取得します。(x$v\link{time == pos})}
}
\description{
setVals関数で作成したデータから、任意の時間位置のデータを取得します。
}
\examples{
vals <- setVals("test1", rnorm(100), 2, "test2", matrix(rnorm(100), ncol = 10), 4.1)
getVals(vals, timelength = 10, pos = 2)
getVals(vals, time = sample(1:10, 100, replace = TRUE), pos = 2)
}
\keyword{internal}
