% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstdbscan.R
\name{hstdbscan}
\alias{hstdbscan}
\title{calc Hierarchical ST-DBSCAN clustering.}
\usage{
hstdbscan(
  x,
  ...,
  dist_option = list(metric = "gower"),
  hclust_option = list(method = "ward.D2")
)
}
\arguments{
\item{x}{位置情報データをsf, sfc, matrix, data.frameのい何れかで指定します。
matrix, data.frame で指定する場合、経度をx, 緯度をyとして列名 (colnames, names) 、データを作成してください。
また、matrix, data.frameで指定した場合は、内部でsfcクラスのデータに変換されます。}

\item{...}{\code{stdbscan} 関数の \code{x} 引数以外の引数を指定します。}

\item{dist_option}{階層的クラスタリングの距離の計算条件を指定します。ここでは \code{cluster::daisy} 関数の引数を \code{dist_option} 引数に \code{list} 形式で指定します。}

\item{hclust_option}{クラスタリングの条件を指定します。 \code{base::hclust} 関数で利用できる引数を \code{list} 形式で指定します。}
}
\value{
\code{stdbscan} の結果に加え、以下の項目が表含まれています。
\itemize{
\item dist: 距離の計算結果テーブルが格納されています。
\item hc: クラスタリングのツリー情報が含まれています。
}
}
\description{
階層的 (Hierarchical) ST-DBSCAN法でクラスタリングを行います。
}
\examples{
eps1 <- 144 ## km
eps2 <- 3600 * 6
time <- NULL
minPts <- 4
FUN <- mean
time <- as.POSIXct("2024-01-01 00:00:00", tz = "JST") + 3600 * seq(0, 23, by = 6)
x <- sf::st_as_sf(expand.grid(seq(130, 140, by = 1), seq(30, 40, by = 1)),
                  coords = c("Var1", "Var2"), crs = 4326) |>
  sf::st_geometry()
vals <- setVals("test1", rnorm(length(x) * length(time), mean = 10, sd = 1), 0.5,
                "test2", matrix(rnorm(length(x) * length(time), mean = 100, sd = 10), ncol = length(time)), 5)
clust <- hstdbscan(x = x, time = time, eps1 = eps1, eps2 = eps2, minPts = 4,
                   vals = vals, neighbortype = "spatial")
cutclust(clust, 4)
}
