% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hstdbscan}
\alias{plot.hstdbscan}
\title{plot cluster dendrogram}
\usage{
\method{plot}{hstdbscan}(x, ...)
}
\arguments{
\item{x}{\code{hstdbscan} 関数の結果オブジェクトを指定します。}

\item{...}{\code{plot} 関数で指定できる引数を指定します。}
}
\description{
plot cluster dendrogram
}
