% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stclust}
\alias{plot.stclust}
\title{plot stclust data}
\usage{
\method{plot}{stclust}(x, bgmap = NULL, ...)
}
\arguments{
\item{x}{\code{stclust} 関数の結果オブジェクトを指定します。}

\item{bgmap}{背景用の地図を sf クラスのデータで指定します。}

\item{...}{\code{plot} 関数で指定できる引数を指定します。}
}
\description{
plot stclust data
}
