% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stdbscan}
\alias{plot.stdbscan}
\title{plot ST-DBSCAN}
\usage{
\method{plot}{stdbscan}(x, ...)
}
\arguments{
\item{x}{\code{stdbscan} 関数の結果オブジェクトを指定します。}

\item{...}{\code{plot} 関数で指定できる引数を指定します。}
}
\description{
\code{stdbscan} 関数で計算した結果オブジェクトを表示します。
}
