% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{print.raw_stdbscan}
\alias{print.raw_stdbscan}
\title{print raw stdbscan data}
\usage{
\method{print}{raw_stdbscan}(x)
}
\arguments{
\item{x}{\code{setOneVals() } 関数で作成したデータを指定します。}
}
\description{
print raw stdbscan data
}
\examples{
setOneVals(name = "ox", v = abs(runif(100)) * 100, deps = 2)
}
