% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.stclust}
\alias{print.stclust}
\title{print stclust class}
\usage{
\method{print}{stclust}(x, ...)
}
\arguments{
\item{x}{"stclust" 関数の結果オブジェクトを指定します。}

\item{...}{\code{print} 関数で使用できる引数を指定します。}
}
\description{
print stclust class
}
