% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rect_hstdbscan}
\alias{rect_hstdbscan}
\title{Draw rectangle using Hierarchical ST-DBSCAN}
\usage{
rect_hstdbscan(x, k, ...)
}
\arguments{
\item{x}{\code{hstdbscan} 関数の結果オブジェクトを指定します。}

\item{k}{クラスター分割数を指定します。}

\item{...}{その他、\code{rect.hclust} で指定できる引数を設定します。}
}
\description{
Draw rectangle using Hierarchical ST-DBSCAN
}
