% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-STclustering.R
\name{rmse}
\alias{rmse}
\title{calc Root-Mean-Squared-Error (RMSE)}
\usage{
rmse(x = double(), y = double())
}
\arguments{
\item{x}{観測値を数値ベクトルで指定します。}

\item{y}{予測値を数値ベクトルで指定します。}
}
\description{
精度評価用に観測値と予測値との比較を行います。
}
\examples{
rmse(x = 1:10,
     y = rnorm(10))
}
