% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{rneighbors}
\alias{rneighbors}
\title{retrive neighbors function.}
\usage{
rneighbors(vals, nb, cluster_pos = NULL, FUN)
}
\arguments{
\item{vals}{\code{list(list(D = ..., delta_eps))} のように指定します。}

\item{nb}{隣接点を vector 形式で指定します。}

\item{cluster_pos}{valsにおける同一クラスターへのインデックスを指定します。}

\item{FUN}{集計用の計算関数を指定します。}
}
\description{
隣接点を探します。
}
\note{
eps1 <- 144
eps2 <- 3600 * 6
minPts <- 5
vals <- setVals("test1", rnorm(100), 0.2,
"test2", matrix(rnorm(100), ncol = 10), 0.2)
\subsection{ramdom point and non temporal data.}{

x <- sf::st_as_sf(data.frame(lon = runif(100, min = 130, max = 140),
lat = runif(100, min = 30, max = 40)),
coords = c("lon", "lat"), crs = 4326) |>
sf::st_geometry()
nb <- stnb(x = x, eps1 = eps1/1000, eps2 = eps2, time = NULL, neighbortype = "random")
message("neighbors: ")
nb$id[\link{1}][\link{1}]
message("cluster neighbors: ")
rneighbors(vals = vals, nb = nb$id[\link{1}][\link{1}], cluster_pos = 1, FUN = mean)
}
}
\keyword{internal}
