% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-STclustering.R
\name{rsq}
\alias{rsq}
\title{calc coefficient of determination}
\usage{
rsq(x = double(), y = double())
}
\arguments{
\item{x}{観測値を数値ベクトルで指定します。}

\item{y}{予測値を数値ベクトルで指定します。}
}
\description{
R ^ 2値を計算します。
}
\examples{
rsq(x = 1:10,
   y = rnorm(10))
}
