% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{setOneVals}
\alias{setOneVals}
\title{set stdbscan's value}
\usage{
setOneVals(name = NULL, v = NULL, deps = NULL)
}
\arguments{
\item{name}{データを判別するための名前を文字列で指定します。}

\item{v}{時間・空間軸のデータをベクトル形式で指定します。}

\item{deps}{クラスターを判別するための閾値 (Δeps) を数値で指定します。}
}
\value{
\itemize{
\item v: 時間・空間軸のデータ
\item deps: クラスターを判別するための閾値 (Δeps)
attributes
\item name: データ名
}
}
\description{
時間・空間軸に該当するデータと、このデータのクラスターを判別するための閾値 (Δeps) を指定します。
}
\examples{
setOneVals(name = "ox", v = abs(runif(100)) * 100, deps = 2)
}
