% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{setVals}
\alias{setVals}
\title{set some stdbscan's value}
\usage{
setVals(...)
}
\arguments{
\item{...}{\code{setOneVlas()} で指定する name, v, deps 引数の値を順番に指定します。
複数のデータを指定する場合は、name, v, deps, name, v, deps, ...のように順番に指定します。}
}
\description{
複数の時間・空間軸に該当するデータと、このデータのクラスターを判別するための閾値 (Δeps) を作成します。
}
\examples{
setVals("test1", rnorm(100), 2, "test2", matrix(rnorm(100), ncol = 10), 4.1)
}
