% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{st_dbscan}
\alias{st_dbscan}
\title{calc ST-DBSCAN clustering and return cluster label.}
\usage{
st_dbscan(
  nb = NULL,
  vals = list(list(D = NULL, delta_eps = 5)),
  type = c("grid", "default"),
  minPts = 10,
  FUN = mean,
  na_value = -9999
)
}
\arguments{
\item{nb}{neighbor list}

\item{vals}{as delta_eps list}

\item{type}{\itemize{
\item "grid": For Grid data. Adjust nb size before calc.
\item "default": For normal(random point) data. not Adjustment.
}}

\item{minPts}{minimum number of points}

\item{FUN}{集計用の計算関数を指定します。}

\item{na_value}{元データ (D) にNA値が含まれている場合に代替する値を指定します。
なお、データは代替して計算されますが、NA値のクラスタリング結果は Noise 判定となります。
Noise判定にしたくない場合はこの機能を利用せず、引数指定時にNA値を除去または代替したデータを設定するようにしてください。}
}
\description{
ST-DBSCAN法でクラスタリングを行います。
https://www.sciencedirect.com/science/article/pii/S0169023X06000218
}
\details{
article{birant2007st,
title={ST-DBSCAN: An algorithm for clustering spatial--temporal data},
author={Birant, Derya and Kut, Alp},
journal={Data \& knowledge engineering},
volume={60},
number={1},
pages={208--221},
year={2007},
publisher={Elsevier}
}
}
