% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstdbscan.R, R/plot.R, R/print.R,
%   R/stdbscan-package.R, R/stdbscan.R
\docType{package}
\name{stdbscan-package}
\alias{stdbscan}
\alias{stdbscan-package}
\title{stdbscan: Spatio Temporal Density Based Scanning Clustering. (One Line, Title Case)}
\description{
What the package does (one paragraph).

What the package does (one paragraph).

What the package does (one paragraph).

What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shoei Furusawa \email{ieohs-furu30@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
