% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{tnearneigh}
\alias{tnearneigh}
\title{Calculate adjacent points of time information}
\usage{
tnearneigh(x, t1 = 0, t2)
}
\arguments{
\item{x}{sfcクラスのデータを指定します。}

\item{t1}{隣接点の最低値を指定します。}

\item{t2}{隣接点の最大値を指定します。}
}
\description{
時間情報ベクトルの隣接情報を作成します。
}
\examples{
td <- sample(seq(as.Date("2025-01-01"), as.Date("2025-02-01"), by = "1 day"), 100, replace = TRUE)
tnearneigh(x = td, t1 = 2, t2 = 4)
}
