% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{validPlot}
\alias{validPlot}
\title{plot for time-series data about sparseDFM model, and check models.}
\usage{
validPlot(x, time_label = seq(1, ncol(x$data)), ...)
}
\arguments{
\item{x}{stclust クラスのデータオブジェクトを指定します。}

\item{time_label}{時系列情報を指定します。}

\item{...}{geom_path で指定できる引数値を指定します。}
}
\description{
plot for time-series data about sparseDFM model, and check models.
}
