% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstdbscan.R
\name{cutclust}
\alias{cutclust}
\title{\code{hstdbscan} 関数で計算したクラスターの結果を分割します。}
\usage{
cutclust(x, k, ...)
}
\arguments{
\item{x}{\code{hstdbscan} の結果オブジェクトを指定します。}

\item{k}{クラスターの分割数を整数で指定します。}

\item{...}{その他、 \code{cutree} 関数で使用できる引数を指定することができます。}
}
\value{
階層的クラスタリング結果のラベル、位置情報が含まれる data.frame が返ります。
}
\description{
\code{hstdbscan} 関数で計算したクラスターの結果を分割します。
}
\examples{
x <- seq(130, 140, by = 1)
y <- seq(30, 40, by = 1)
t <- as.POSIXct("2024-01-01 00:00:00", tz = "JST") + 3600 * seq(0, 23, by = 6)
geo <- sf::st_as_sf(expand.grid(x, y), coords = c("Var1", "Var2"), crs = 4326)

D <- abs(runif(nrow(geo) * length(t))) * 100
clust <- hstdbscan(x = geo, time = t, eps1 = 144, eps2 = 3600 * 6, minPts = 6,
                  vals = list(list(D = D,
                                   delta_eps = 20)),
                  metric = "geo", neighbortype = "spatial", dbscantype = "grid")
cutclust(clust, 4)
}
