% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstdbscan.R
\name{hstdbscan}
\alias{hstdbscan}
\title{calc Hierarchical ST-DBSCAN clustering.}
\usage{
hstdbscan(
  x,
  ...,
  dist_option = list(metric = "gower"),
  hclust_option = list(method = "ward.D2")
)
}
\arguments{
\item{x}{\code{stdbscan} 関数の \code{x} 引数の説明を参照してください。}

\item{...}{\code{stdbscan} 関数の \code{x} 引数以外の引数を指定します。}

\item{dist_option}{階層的クラスタリングの距離の計算条件を指定します。ここでは \code{cluster::daisy} 関数の引数を \code{dist_option} 引数に \code{list} 形式で指定します。}

\item{hclust_option}{クラスタリングの条件を指定します。 \code{base::hclust} 関数で利用できる引数を \code{list} 形式で指定します。}
}
\value{
\code{stdbscan} の結果に加え、以下の項目が表含まれています。
\itemize{
\item dist: 距離の計算結果テーブルが格納されています。
\item hc: クラスタリングのツリー情報が含まれています。
}
}
\description{
階層的 (Hierarchical) ST-DBSCAN法でクラスタリングを行います。
}
\examples{
x <- seq(130, 140, by = 1)
y <- seq(30, 40, by = 1)
t <- as.POSIXct("2024-01-01 00:00:00", tz = "JST") + 3600 * seq(0, 23, by = 6)
geo <- sf::st_as_sf(expand.grid(x, y), coords = c("Var1", "Var2"), crs = 4326)

D <- abs(runif(nrow(geo) * length(t))) * 100
clust <- hstdbscan(x = geo, time = t, eps1 = 144, eps2 = 3600 * 6, minPts = 6,
                  vals = list(list(D = D,
                                   delta_eps = 20)),
                  metric = "geo", neighbortype = "spatial", dbscantype = "grid")
cutclust(clust, 4)
}
