% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{stnb}
\alias{stnb}
\title{create neighbor list}
\usage{
stnb(
  x = double(),
  y = double(),
  time = NULL,
  eps0 = 0,
  eps1 = NULL,
  eps2 = NULL,
  method = c("geo", "euclidean"),
  neighbortype = c("spatial", "random")
)
}
\arguments{
\item{x}{double(), sf, sfc}

\item{y}{double()}

\item{time}{numeric, date, ISOdate, POSIXct}

\item{eps0}{dist x and y (lower)}

\item{eps1}{dist x and y (uppwer)}

\item{eps2}{dist time. set seconds.}

\item{method}{use spdep::dnearneigh.
\itemize{
\item "geo": calc for geographics.
\item "euclidean": calc for euclidean area.
}}

\item{neighbortype}{\itemize{
\item "Spatial": For Grid data. Adjust nb size before calc.
\item "Random": For normal(random point) data. not Adjustment.
}}
}
\description{
隣接点行列を作成します。
}
